requirements.txt
numpy
pandas
python-dotenv
mysql-connector-python
pymysql
scikit-learn
#-e .           ## aa setup.py install , ne direct call kare